Refactoring; major feature parity in place
