+++ gooey/gui/components/util/wrapped_static_text.py
@@ -44,7 +44,6 @@ class AutoWrappedStaticText(wx.StaticText):
         """
 
         event.Skip()
-        print(self.GetLabel())
         if self.target:
             self.Wrap(self.target.GetSize().width)
         else:
