+++ gooey/gui/components/widgets/textarea.py
@@ -7,6 +7,7 @@ from gooey.gui.components.widgets.textfield import TextField
 from gooey.gui.components.widgets.bases import TextContainer
 from gooey.gui import formatters
 from gooey.python_bindings import types as t
+from python_bindings.types import FormField
 
 
 class Textarea(TextContainer):
@@ -37,5 +38,18 @@ class Textarea(TextContainer):
     def formatOutput(self, metatdata, value: str):
         return formatters.general(metatdata, value.replace('\n', os.linesep))
 
+    def syncUiState(self, state: FormField):
+        self.setValue(state['value'])
+        self.error.SetLabel(state['error'] or '')
+        self.error.Show(state['error'] is not None and state['error'] is not '')
+
     def getUiState(self) -> t.FormField:
-        return t.Textarea(**super().getUiState())
+        return t.TextField(
+            id=self._id,
+            type=self.widgetInfo['type'],
+            value=self.getWidgetValue(),
+            placeholder=self.widget.GetHint(),
+            error=self.error.GetLabel().replace('\n', ' '),
+            enabled=self.IsEnabled(),
+            visible=self.IsShown()
+        )
