+++ gooey/gui/components/widgets/textfield.py
@@ -26,5 +26,6 @@ class TextField(TextContainer):
         textctr.SetHint(state['placeholder'])
         textctr.Enable(state['enabled'])
         self.Show(state['visible'])
-        if state['error']:
-            self.setErrorString(state['error'])
+        self.error.SetLabel(state['error'] or '')
+        self.error.Show(state['error'] is not None and state['error'] is not '')
+        self.Layout()
