+++ gooey/gui/containers/application.py
@@ -440,595 +440,7 @@ class GooeyApplication(wx.Frame):
         self.header.setSubtitle(_('finished_forced_quit'))
 
 
-class HeaderProps(TypedDict):
-    background_color: str
-    title: str
-    show_title: bool
-    subtitle: str
-    show_subtitle: bool
 
 
-class RFooter(Component):
-    def __init__(self, props):
-        super().__init__(props)
-        self.ref = Ref()
 
-    def component_did_mount(self):
-        """
-        We have to manually wire up LEFT_DOWN handlers
-        for every component due to wx limitations.
-        See: mouse.py docs for background.
-        """
-        block: wx.BoxSizer = self.ref.instance
-        for child in block.Children:
-            child.Bind(wx.EVT_LEFT_DOWN, notifyMouseEvent)
 
-
-    def handle(self, btn):
-        def inner(*args, **kwargs):
-            pub.send_message(btn['id'])
-        return inner
-
-
-    def render(self):
-        return wsx(
-            [c.Block, {'orient': wx.VERTICAL,
-                       'min_size': (30, 53),
-                       'background_color': self.props['bg_color']},
-             [c.Block, {'orient': wx.VERTICAL, 'proportion': 1}],
-             [c.Block, {'orient': wx.HORIZONTAL,
-                        'border': 20,
-                        'flag': wx.EXPAND | wx.LEFT | wx.RIGHT,
-                        'ref': self.ref},
-              [c.Gauge, {'range': 100,
-                         'proportion': 1,
-                         'value': self.props['progress']['value'],
-                         'show': self.props['progress']['show']}],
-              [c.StaticText, {'label': present_time(self.props['timing']),
-                              'flag': wx.LEFT,
-                              # TODO: pass independent Show prop
-                              'show': self.props['progress']['show'],
-                              'border': 20}],
-              [c.Block, {'orient': wx.HORIZONTAL, 'proportion': 1}],
-              *[[c.Button, {**btn,
-                            'label': _(btn['label_id']),
-                            'min_size': (90, 23),
-                            'flag': wx.LEFT,
-                            'border': 10,
-                            'on_click': self.handle(btn)
-                            }]
-                for btn in self.props['buttons']]],
-             [c.Block, {'orient': wx.VERTICAL, 'proportion': 1}]]
-        )
-
-
-
-# @mount.register(Tabbar)
-# def tabbar(element, parent):
-#     return update(element, Tabbar(parent, xxx, {'contents': []}))
-#
-#
-# @update.register(Tabbar)
-# def tabbar(element, instance: Tabbar):
-#     set_basic_props(instance, element['props'])
-#     return instance
-
-
-# @mount.register(Sidebar)
-# def sidebar(element, parent):
-#     return update(element, Sidebar(parent, xxx, {'contents': []}))
-#
-#
-# @update.register(Sidebar)
-# def sidebar(element, instance: Sidebar):
-#     set_basic_props(instance, element['props'])
-#     return instance
-
-
-@mount.register(Console)
-def console(element, parent):
-    return update(element, Console(parent, element['props']))
-
-
-@update.register(Console)
-def console(element, instance: Console):
-    set_basic_props(instance, element['props'])
-    if 'show' in element['props']:
-        instance.Show(element['props']['show'])
-    return instance
-
-
-
-class RNavbar(Component):
-    def __init__(self, props):
-        super().__init__(props)
-
-    # if self.buildSpec['navigation'] == constants.TABBED:
-    #     navigation = Tabbar(self, self.buildSpec, self.configs)
-    # else:
-    #     navigation = Sidebar(self, self.buildSpec, self.configs)
-    #     if self.buildSpec['navigation'] == constants.HIDDEN:
-    #         navigation.Hide()
-    def render(self):
-        return wsx(
-
-        )
-
-def VerticalSpacer(props):
-    return wsx([c.Block, {'orient': wx.VERTICAL, 'min_size': (-1, props['height'])}])
-
-def SidebarControls(props):
-    return wsx(
-        [c.Block, {'orient': wx.VERTICAL,
-                   'min_size': (180, 0),
-                   'size': (180, 0),
-                   'show': props.get('show', True),
-                   'flag': wx.EXPAND,
-                   'proportion': 0,
-                   'background_color': props['bg_color']},
-         [c.Block, {'orient': wx.VERTICAL,
-                    'min_size': (180, 0),
-                    'size': (180, 0),
-                    'flag': wx.EXPAND | wx.LEFT | wx.RIGHT | wx.BOTTOM,
-                    'border': 10,
-                    'proportion': 1,
-                    'background_color': props['bg_color']},
-          [VerticalSpacer, {'height': 15}],
-          [TitleText, {'label': props['label']}],
-          [VerticalSpacer, {'height': 5}],
-          [c.ListBox, {'choices': props['options'],
-                       'value': props['activeSelection'],
-                       'proportion': 1,
-                       'on_change': props['on_change'],
-                       'flag': wx.EXPAND}],
-          [VerticalSpacer, {'height': 10}]]]
-    )
-
-
-def ProgressSpinner(props):
-    return wsx(
-        [c.Block, {'flag': wx.EXPAND, 'show': props['show']},
-         [c.Gauge, {'flag': wx.EXPAND,
-                    'value': -1,
-                    'size': (-1, 4)}],
-         [c.StaticLine, {'style': wx.LI_HORIZONTAL,
-                         'flag': wx.EXPAND}]]
-    )
-
-
-def ErrorWarning(props):
-    return wsx(
-        [c.Block, {'orient': wx.HORIZONTAL,
-                   'background_color': '#fdeded',
-                   'style': wx.SIMPLE_BORDER,
-                   'flag': wx.EXPAND | wx.ALL,
-                   'proportion': 0,
-                   'border': 5,
-                   'min_size': (-1, 45),
-                   'show': props.get('show', True)},
-         [c.StaticBitmap, {'size': (24, 24),
-                           'flag': wx.LEFT | wx.RIGHT | wx.ALIGN_CENTER_VERTICAL,
-                           'border': 6,
-                           'uri': props['uri']}],
-         [c.StaticText, {'label': 'Whoops! You have some errors which must be corrected',
-                         'flag': wx.ALIGN_CENTER_VERTICAL}]]
-    )
-
-def RSidebar(props):
-    return wsx(
-        [c.Block, {'orient': wx.HORIZONTAL,
-                   'show': props.get('show', True),
-                   'flag': props['flag'],
-                   'proportion': props['proportion'],
-                   'ref': props['ref']},
-         [SidebarControls, props],
-         [c.StaticLine, {'style': wx.LI_VERTICAL,
-                         'flag': wx.EXPAND,
-                         'min_size': (1, -1)}],
-         *[[ConfigPage, {'flag': wx.EXPAND,
-                         'proportion': 3,
-                         'config': config,
-                         'show': i == props['activeSelection']}]
-           for i, config in enumerate(props['config'].values())]
-         ]
-    )
-
-
-def TabbedForm(props):
-    return wsx(
-        [c.Notebook, {'flag': wx.EXPAND, 'proportion': 1, 'on_change': props['on_change']},
-         [c.NotebookItem, {'title': 'Page 1', 'selected': props['activeTab'] == 0},
-          [ConfigPage, {'flag': wx.EXPAND, 'proportion': 1}]],
-         [c.NotebookItem, {'title': 'Page 2!!!', 'selected': props['activeTab'] == 1},
-          [ConfigPage, {'flag': wx.EXPAND, 'proportion': 1}]]],
-    )
-
-
-class RGooey(Component):
-    def __init__(self, props):
-        super().__init__(props)
-        self.frameRef = Ref()
-        self.consoleRef = Ref()
-        self.configRef = Ref()
-
-        self.buildSpec = props
-        self.state = initial_state(props)
-        self.headerprops = lambda state: {
-            'background_color': self.buildSpec['header_bg_color'],
-            'title': state['title'],
-            'subtitle': state['subtitle'],
-            'flag': wx.EXPAND,
-            'height': self.buildSpec['header_height'],
-            'image_uri': state['image'],
-            'image_size': (six.MAXSIZE, self.buildSpec['header_height'] - 10)}
-
-        self.fprops = lambda state: {
-            'buttons': state['buttons'],
-            'progress': state['progress'],
-            'timing': state['timing'],
-            'bg_color': self.buildSpec['footer_bg_color'],
-            'flag': wx.EXPAND,
-        }
-        self.clientRunner = ProcessController.of(self.buildSpec)
-
-
-    def component_did_mount(self):
-        pub.subscribe(events.WINDOW_START, self.onStart)
-        pub.subscribe(events.WINDOW_RESTART, self.onStart)
-        pub.subscribe(events.WINDOW_STOP, self.handleInterrupt)
-        pub.subscribe(events.WINDOW_CLOSE, self.handleClose)
-        pub.subscribe(events.WINDOW_CANCEL, self.handleCancel)
-        pub.subscribe(events.WINDOW_EDIT, self.handleEdit)
-        pub.subscribe(events.CONSOLE_UPDATE, self.consoleRef.instance.logOutput)
-        pub.subscribe(events.EXECUTION_COMPLETE, self.handleComplete)
-        pub.subscribe(events.PROGRESS_UPDATE, self.updateProgressBar)
-        pub.subscribe(events.TIME_UPDATE, self.updateTime)
-        # # Top level wx close event
-        self.frameRef.instance.Bind(wx.EVT_CLOSE, self.handleClose)
-        self.frameRef.instance.SetMenuBar(MenuBar(self.buildSpec))
-
-    def getActiveConfig(self):
-        return [config
-                for config in self.configRef.instance.Children
-                if isinstance(config, ConfigPage)][self.state['activeSelection']]
-
-    def getActiveFormState(self):
-        """
-        This boiler-plate and manual interrogation of the UIs
-        state is required until we finish porting the Config Form
-        over to rewx (which is a battle left for another day given
-        its complexity)
-        """
-        return self.getActiveConfig().getFormState()
-
-
-    def fullState(self):
-        """
-        Re: final porting is a to do. For now we merge the UI
-        state into the main tracked state.
-        """
-        formState = self.getActiveFormState()
-        return s.combine(self.state, self.props, formState)
-
-    def onStart(self, *args, **kwargs):
-        if Events.VALIDATE_FORM in self.state['use_events']:
-            self.runAsyncValidation()
-        else:
-            self.startRun()
-
-    def startRun(self):
-        state = self.fullState()
-        if state['clear_before_run']:
-            self.consoleRef.instance.Clear()
-        self.set_state(s.consoleScreen(_, state))
-        self.clientRunner.run(s.build_cli(state))
-        self.frameRef.instance.Layout()
-        for child in self.frameRef.instance.Children:
-            child.Layout()
-
-
-    def syncExternalState(self, state: FullGooeyState):
-        """
-        Sync the UI's state to what the host program has requested.
-        """
-        self.getActiveConfig().syncFormState(s.activeFormState(state))
-        self.frameRef.instance.Layout()
-        for child in self.frameRef.instance.Children:
-            child.Layout()
-
-
-    def runAsyncValidation(self):
-        def handleHostResponse(validationErrors: Dict[str, str]):
-            self.set_state(s.finishUpdate(self.state))
-            if validationErrors:
-                currentState = self.fullState()
-                nextState = s.updateErrors(currentState, validationErrors)
-                self.syncExternalState(nextState)
-            else:
-                print("Calling start run from thread ", threading.get_ident())
-                self.startRun()
-
-        def onComplete(result: Try[Dict[str, str]]):
-            result.onSuccess(handleHostResponse)
-            result.onError(self.handleHostError)
-
-        print("Hello from thread", get_ident())
-        self.set_state(s.beginUpdate(self.state))
-        fullState = self.fullState()
-        host.communicateFormValidation(fullState, callafter(onComplete))
-
-
-    def runAsyncExternalOnCompleteHandler(self, was_success):
-        def handleHostResponse(hostState):
-            try:
-                if hostState:
-                    currentState = self.fullState()
-                    self.syncExternalState(s.mergeExternalState(currentState, hostState))
-                else:
-                    print("Calling start run from thread ", threading.get_ident())
-                    wx.CallAfter(self.startRun)
-            except Exception as e:
-                print(e)
-                print('uh oh!')
-
-        def onComplete(result: Try[PublicGooeyState]):
-            self.set_state({**self.state, 'fetchingUpdate': False})
-            result.onError(self.handleHostError)
-            result.onSuccess(handleHostResponse)
-
-        print("Hello from thread", get_ident())
-        fullState = self.fullState()
-        if was_success:
-            host.communicateSuccessState(fullState, callafter(onComplete))
-        else:
-            host.communicateErrorState(fullState, callafter(onComplete))
-
-
-    def handleHostError(self, exception):
-        try:
-            raise exception
-        except CalledProcessError as e:
-            self.set_state(s.errorScreen(_, self.state))
-            self.consoleRef.instance.appendText(str(e))
-            self.consoleRef.instance.appendText(
-                f'\n\nThis failure happens when Gooey tries to invoke your '
-                'code for the TODO event and receives an unexpected '
-                'error code in response')
-            wx.CallAfter(modals.showFailure)
-        except JSONDecodeError as e:
-            self.set_state(s.errorScreen(_, self.state))
-            self.consoleRef.instance.appendText(str(e))
-            self.consoleRef.instance.appendText(
-                f'\n\nGooey was unable to parse the response to the TODO event. '
-                'This can happen if you have additional logs to stdout beyond what Gooey '
-                'expects.')
-            wx.CallAfter(modals.showFailure)
-        except Exception as e:
-            self.set_state(s.errorScreen(_, self.state))
-            self.consoleRef.instance.appendText(str(e))
-        finally:
-            self.set_state({**self.state, 'fetchingUpdate': False})
-
-
-    def handleInterrupt(self, *args, **kwargs):
-        if self.shouldStopExecution():
-            self.clientRunner.stop()
-
-
-    def handleComplete(self, *args, **kwargs):
-        if self.clientRunner.was_success():
-            self.handleSuccessfulRun()
-            self.runAsyncExternalOnCompleteHandler(was_success=True)
-        else:
-            self.handleErrantRun()
-            self.runAsyncExternalOnCompleteHandler(was_success=False)
-
-    def handleSuccessfulRun(self):
-        if self.state['return_to_config']:
-            self.set_state(s.editScreen(_, self.state))
-        else:
-            self.set_state(s.successScreen(_, self.state))
-            if self.state['show_success_modal']:
-                wx.CallAfter(modals.showSuccess)
-
-
-    def handleErrantRun(self):
-        # with self.handleHostError(Events.ON_ERROR):
-        #     updates = host.fetchOnErrorState()
-        #     self.getActiveConfig().syncFormState(updates)
-
-        if self.clientRunner.wasForcefullyStopped:
-            self.set_state(s.interruptedScreen(_, self.state))
-        else:
-            self.set_state(s.errorScreen(_, self.state))
-            if self.state['show_failure_modal']:
-                wx.CallAfter(modals.showFailure)
-
-
-    def successScreen(self):
-        strings = {'title': _('finished_title'), 'subtitle': _('finished_msg')}
-        self.set_state(s.success(self.state, strings, self.buildSpec))
-
-
-    def showForceStopped(self):
-        self.showComplete()
-        if self.buildSpec.get('force_stop_is_error', True):
-            self.showError()
-        else:
-            self.showSuccess()
-        self.header.setSubtitle(_('finished_forced_quit'))
-
-    def onCompleteOLD(self, *args, **kwargs):
-        """
-        Display the appropriate screen based on the success/fail of the
-        host program
-        """
-        with transactUI(self):
-            if self.clientRunner.was_success():
-                if self.buildSpec.get('return_to_config', False):
-                    self.showSettings()
-                else:
-                    self.showSuccess()
-                    if self.buildSpec.get('show_success_modal', True):
-                        wx.CallAfter(modals.showSuccess)
-            else:
-                if self.clientRunner.wasForcefullyStopped:
-                    self.showForceStopped()
-                else:
-                    self.showError()
-                    if self.buildSpec.get('show_failure_modal'):
-                        wx.CallAfter(modals.showFailure)
-
-    def handleEdit(self, *args, **kwargs):
-        self.set_state(s.editScreen(_, self.state))
-
-
-    def handleCancel(self, *args, **kwargs):
-        if modals.confirmExit():
-            self.handleClose()
-
-    def handleClose(self, *args, **kwargs):
-        """Stop any actively running client program, cleanup the top
-        level WxFrame and shutdown the current process"""
-        # issue #592 - we need to run the same onStopExecution machinery
-        # when the exit button is clicked to ensure everything is cleaned
-        # up correctly.
-        frame: wx.Frame = self.frameRef.instance
-        if self.clientRunner.running():
-            if self.shouldStopExecution():
-                self.clientRunner.stop()
-                frame.Destroy()
-                # TODO: NOT exiting here would allow
-                # spawing the gooey to input params then
-                # returning control to the CLI
-                sys.exit()
-        else:
-            frame.Destroy()
-            sys.exit()
-
-    def shouldStopExecution(self):
-        return not self.state['show_stop_warning'] or modals.confirmForceStop()
-
-    def updateProgressBar(self, *args, **kwargs):
-        self.set_state(s.updateProgress(self.state, ProgressEvent(**kwargs)))
-
-    def updateTime(self, *args, **kwargs):
-        self.set_state(s.updateTime(self.state, TimingEvent(**kwargs)))
-
-    def handleSelectAction(self, event):
-        self.set_state(assoc(self.state, 'activeSelection', event.Selection))
-
-
-    def render(self):
-        return wsx(
-            [c.Frame, {'title': self.buildSpec['program_name'],
-                       'background_color': self.buildSpec['body_bg_color'],
-                       'double_buffered': True,
-                       'min_size': (400, 300),
-                       'size': self.buildSpec['default_size'],
-                       'ref': self.frameRef},
-             [c.Block, {'orient': wx.VERTICAL},
-              [RHeader, self.headerprops(self.state)],
-              [c.StaticLine, {'style': wx.LI_HORIZONTAL, 'flag': wx.EXPAND}],
-              [ProgressSpinner, {'show': self.state['fetchingUpdate']}],
-              [ErrorWarning, {'show': False,
-                              'uri': self.state['images']['errorIcon']}],
-              [Console, {**self.buildSpec,
-                         'flag': wx.EXPAND,
-                         'proportion': 1,
-                         'show': self.state['screen'] == 'CONSOLE',
-                         'ref': self.consoleRef}],
-              [RSidebar, {'bg_color': self.buildSpec['sidebar_bg_color'],
-                          'label': 'Some Action!',
-                          'ref': self.configRef,
-                          'show': self.state['screen'] == 'FORM',
-                          'activeSelection': self.state['activeSelection'],
-                          'options': list(self.buildSpec['widgets'].keys()),
-                          'on_change': self.handleSelectAction,
-                          'config': self.buildSpec['widgets'],
-                          'flag': wx.EXPAND,
-                          'proportion': 1}],
-              # [c.Notebook, {'flag': wx.EXPAND | wx.ALL, 'border': 20, 'proportion': 1, 'on_change': lambda *args: None},
-              #  [c.NotebookItem, {'title': 'Page 1', 'selected': self.state['activeSelection'] == 0},
-              #   [c.StaticText, {'label': "Hello!"}]],
-              #  [c.NotebookItem, {'title': 'Page 2!!!', 'selected': self.state['activeSelection'] == 1},
-              #   [c.StaticText, {'label': "Hello!"}]]],
-              [c.StaticLine, {'style': wx.LI_HORIZONTAL, 'flag': wx.EXPAND}],
-              [RFooter, self.fprops(self.state)]]]
-        )
-
-
-class RHeader(Component):
-    def __init__(self, props):
-        super().__init__(props)
-
-    def render(self):
-        ps = {
-                 'uri': self.props['image_uri'],
-                 'size': self.props['image_size'],
-                 'flag': wx.RIGHT,
-                 'border': 10}\
-            if 'running' not in self.props['image_uri']\
-            else {'size': self.props['image_size'],
-                 'flag': wx.RIGHT,
-                 'border': 10}
-        return wsx(
-            [c.Block, {'orient': wx.HORIZONTAL,
-                       'min_size': (120, self.props['height']),
-                       'background_color': self.props['background_color']},
-             [c.Block, {'orient': wx.VERTICAL,
-                        'flag': wx.ALIGN_CENTER_VERTICAL | wx.ALL,
-                        'proportion': 1,
-                        'border': 10},
-
-              [TitleText, {'label': self.props['title']}],
-              [c.StaticText, {'label': self.props['subtitle']}]],
-             [c.StaticBitmap, ps]]
-        )
-
-
-
-class TitleText(Component):
-    def __init__(self, props):
-        super().__init__(props)
-        self.ref = Ref()
-
-    def component_did_mount(self):
-        text: wx.StaticText = self.ref.instance
-        font_size = text.GetFont().GetPointSize()
-        text.SetFont(wx.Font(
-            int(font_size * 1.2),
-            wx.FONTFAMILY_DEFAULT,
-            Constants.WX_FONTSTYLE_NORMAL,
-            wx.FONTWEIGHT_BOLD,
-            False
-        ))
-
-    def render(self):
-        return wsx([c.StaticText, {'label': self.props['label'], 'ref': self.ref}])
-
-
-
-
-
-@mount.register(ConfigPage)
-def config(element, parent):
-    return update(element, ConfigPage(parent, element['props']['config'], {'contents': []}))
-
-
-@update.register(ConfigPage)
-def config(element, instance: ConfigPage):
-    set_basic_props(instance, element['props'])
-    return instance
-
-
-
-def callafter(f):
-    """
-    Wraps the supplied function in a wx.CallAfter
-    for Thread-safe interop with WX.
-    """
-    @wraps(f)
-    def inner(*args, **kwargs):
-        wx.CallAfter(f, *args, **kwargs)
-    return inner
