+++ gooey/python_bindings/parameters.py
@@ -133,8 +133,19 @@ def gooey_params(**kwargs) -> GooeyParams:
 
 
 def parse_events(events: List[str]) -> List[str]:
+    if not isinstance(events, list):
+        raise TypeError(
+            f"use_events requires a list of events. You provided "
+            "{events}. \n"
+            "Example: \n"
+            "\tfrom gooey import Events"
+            "\t@Gooey(use_events=[Events.VALIDATE_FORM]")
+
     unknown_events = set(events) - set(Events)
     if unknown_events:
-        raise ValueError('Unrecognized events: ', unknown_events)
+        raise ValueError(
+            f'Unrecognized event(s): {unknown_events}\n'
+            f'Must be one of {Events._fields}\n'
+            f'Consider using the `Events` object: `from gooey import Events`')
     else:
         return events
